+++ gateway/src/downstream/mod.rs
@@ -2,5 +2,6 @@ mod extras;
 mod account;
 mod messaging;
 mod channel;
+mod operations;
 
 pub mod downstream;
